# platform-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/device/list": {
            "get": {
                "description": "DeviceList Api 查询用户拥有的设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device"
                ],
                "summary": "DeviceList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页索引",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceListRes"
                        }
                    }
                }
            }
        },
...
        "/api/device_group/create": {
            "put": {
                "description": "DeviceGroupCreate Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device-Group"
                ],
                "summary": "DeviceGroupCreate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create DeviceGroup request body",
                        "name": "DeviceGroupCreate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceGroupCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/device_group/delete": {
            "delete": {
                "description": "DeviceGroupDelete Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device-Group"
                ],
                "summary": "DeviceGroupDelete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Del DeviceGroup request body",
                        "name": "DeviceGroupDelete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceGroupDelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/device_group/detail": {
            "get": {
                "description": "DeviceGroupDetail Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device-Group"
                ],
                "summary": "DeviceGroupDetail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "device_group_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceGroupDetail"
                        }
                    }
                }
            }
        },
        "/api/device_group/list": {
            "get": {
                "description": "DeviceGroupList Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device-Group"
                ],
                "summary": "DeviceGroupList",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页索引",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页大小",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceGroupListRes"
                        }
                    }
                }
            }
        },
        "/api/device_group/update": {
            "post": {
                "description": "DeviceGroupUpdate Api",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Device-Group"
                ],
                "summary": "DeviceGroupUpdate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update DeviceGroup request body",
                        "name": "DeviceGroupUpdate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.DeviceGroupUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.DeviceDetail": {
            "type": "object",
            "properties": {
                "device_name": {
                    "type": "string"
                },
                "device_type": {
                    "$ref": "#/definitions/enum.DeviceType"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "os_type": {
                    "type": "string"
                },
                "status": {
                    "enum": [
                        0,
                        1
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.DeviceStatus"
                        }
                    ]
                }
            }
        },
        "dto.DeviceListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DeviceDetail"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DeviceGroupCreateReq": {
            "type": "object",
            "required": [
                "device_group_name"
            ],
            "properties": {
                "device_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "device_group_name": {
                    "type": "string"
                }
            }
        },
        "dto.DeviceGroupDelReq": {
            "type": "object",
            "required": [
                "device_group_id"
            ],
            "properties": {
                "device_group_id": {
                    "type": "string"
                }
            }
        },
        "dto.DeviceGroupDetail": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Device"
                    }
                },
                "device_group_name": {
                    "type": "string"
                }
            }
        },
        "dto.DeviceGroupListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.DeviceGroupDetail"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DeviceGroupUpdateReq": {
            "type": "object",
            "required": [
                "device_group_id"
            ],
            "properties": {
                "new_device_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "device_group_id": {
                    "type": "integer"
                }
            }
        },
...
        "enum.DeviceStatus": {
            "type": "integer",
            "enum": [
                0,
                1
            ],
            "x-enum-varnames": [
                "DeviceStatusOffline",
                "DeviceStatusOnLine"
            ]
        },
        "enum.DeviceType": {
            "type": "integer",
            "enum": [
                1,
                2
            ],
            "x-enum-varnames": [
                "DeviceTypePhysical",
                "DeviceTypeVirtual"
            ]
        },
...
        "models.Device": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "device_name": {
                    "type": "string"
                },
                "device_type": {
                    "description": "1:实体服务器 2:虚拟机",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.DeviceType"
                        }
                    ]
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "mac": {
                    "type": "string"
                },
                "os_type": {
                    "type": "string"
                },
                "status": {
                    "description": "1:运行中",
                    "allOf": [
                        {
                            "$ref": "#/definitions/enum.DeviceStatus"
                        }
                    ]
                },
                "unique_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}